refactor(worker): split manager lifecycle files
